Feat/display chat response (#5)
* feat: Display chat response on the screen

This commit implements the functionality to display the bot's response on the chat screen after the user sends a message.

The `AiBotScreen` has been converted to a `StatefulWidget` to manage the chat history. New widgets have been created to display user messages and the bot's rich response, which includes a title, total spent, breakdown, spending trend, and summary.

The `ChatInputField` has been modified to use a callback to notify the `AiBotScreen` of new messages. A hardcoded bot response is used to demonstrate the final look and feel of the chat screen.

* made changes in the frontend to display chat messages. pending alignment to UI design.

* Completed end to end flow of fund transfer. Project in working stage.

* Changed the spends output to align with the design. Masked the otp.

---------

Authored-by: Syed Tajamul Shah <syed.tajamul@i-exceed.com>